src/requirements.txt
fastapi[standard]>=0.116.1
llama-index>=0.13.6
llama-index-llms-openai-like>=0.3.0
llama-index-vector-stores-postgres>=0.6.5
httpx>=0.27.0
pgvector>=0.4.1
psycopg>=3.2.9
pydantic>=2.11.7
pydantic-settings>=2.10.1
sqlalchemy>=2.0.43
sqlmodel>=0.0.24
requests>=2.31.0
beautifulsoup4>=4.12.3
pypdf>=3.17.0
datasets
pandas
biopython
supabase
pyiceberg
pytesseract>=0.3.10
Pillow>=10.0.0
arxiv
slowapi>=0.1.9